feat(lib): reorg manifest work into mod
